requirements.txt
# ====================================================================
# Group 21 — Privacy-Preserving Telehealth IoT System
# Unified Python dependencies for every module in this repository.
# ====================================================================

# ----- Web frameworks & HTTP -----
flask>=3.0,<4.0
requests>=2.31

# ----- Validation (ZYM: cloud_schema.py) -----
pydantic>=2.5,<3.0

# ----- Auth (ZYM: cloud_auth.py) -----
PyJWT>=2.7

# ----- Database (YYX: cloud.py) -----
pymongo>=4.6
dnspython>=2.6        # required by MongoDB srv:// URIs

# ----- Cryptography (LYZ: HMAC, key rotation) -----
cryptography>=42.0

# ----- Numerics & ML (ZYM: gateway_dp.py, gateway_ai_ids.py) -----
numpy>=1.24
scikit-learn>=1.3

# ----- Data pipeline (LKK: data cleaning) -----
pandas>=2.0

# ----- Plotting (ZYM & LKK: experiment figures) -----
matplotlib>=3.7

# ----- Dev / test -----
pytest>=7.4
python-dotenv>=1.0    # load .env during local dev
